pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["JuanDRP <juandrp2004@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.11"
langchain = "^0.0.345"
openai = "^1.3.7"
boto3 = "^1.34.39"
pydantic = "^2.6.1"
black = "^24.1.1"
google-api-python-client = "^2.118.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
requests = "^2.31.0"
google-auth = "^2.28.0"
pyjwt = "^2.8.0"
cryptography = "^42.0.3"
pytz = "^2024.1"
pymongo = {extras = ["srv"], version = "^4.6.3"}
pytest = "^8.2.0"
coverage = "^7.5.1"
pre-commit = "^3.7.0"
pytest-cov = "^5.0.0"
twilio = "^6.50.1"
flask = "^3.1.1"
python-dotenv = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
